Add files via upload
This is a small Project, Where it shows occupied and available seats (only non-occupied, seats can be selected). there are movies displayed with Prices, and when selecting seats counter shows the amount. Also clear all button is added because if refreshed local-storage is used to save the changes even after the refresh.